requirements.txt
# SMS Spoofing System Dependencies
flask==2.3.3
flask-cors==4.0.0
requests==2.31.0
python-dotenv==1.0.0
cryptography==41.0.7
pyjwt==2.8.0
sqlalchemy==2.0.23
flask-sqlalchemy==3.1.1
bcrypt==4.0.1
pyserial==3.5
twilio==8.10.0
vonage==3.10.0
clicksend-client==5.0.82
plivo==4.44.0
messagebird==2.2.0
logging==0.4.9.6
schedule==1.2.0
python-dateutil==2.8.2
phonenumbers==8.13.25
validators==0.22.0
psutil==5.9.6
watchdog==3.0.0
